+++ NEWS.md
@@ -30,6 +30,7 @@ Active logins are now necessary to use webservices such as CSI:FingerID, see e.g
     - Fixed: `reportHTML()` could show MS/MS similarity plots of wrong feature groups
     - Fixed: Set specific spectral similarities were not assigned correctly during TP componentization if a feature group occurs multiple times in the same component
 - GenForm: `thrMS`, `thrMSMS`, `thrComb` and `maxCandidates` arguments, which can be used to tweak calculations for features with many candidates, e.g., to limit calculation times.
+- `as.data.table()` for `featureGroups`: added p values when `regression=TRUE`